+++ src/auth/application/usecases/logout.usecase.ts
@@ -1,21 +1,21 @@
 import { Inject, Injectable } from '@nestjs/common';
-import { AUTH_REPOSITORY } from '../../domain/auth.repository';
-import type { AuthRepository } from '../../domain/auth.repository';
 import { AuthPlatform } from '../../domain/auth.types';
+import { AUTH_SESSION_PORT } from '../ports/auth-session.port';
+import type { AuthSessionPort } from '../ports/auth-session.port';
 import { LogoutResult } from '../auth.types';
 
 @Injectable()
 export class LogoutUseCase {
   constructor(
-    @Inject(AUTH_REPOSITORY)
-    private readonly authRepository: AuthRepository,
+    @Inject(AUTH_SESSION_PORT)
+    private readonly authSessionPort: AuthSessionPort,
   ) {}
 
   async execute(
     authAccountId: string,
     platform: AuthPlatform,
   ): Promise<LogoutResult> {
-    await this.authRepository.revokeRefreshTokens(authAccountId, platform);
+    await this.authSessionPort.revokeRefreshTokens(authAccountId, platform);
 
     return { success: true };
   }
