+++ src/auth/application/usecases/refresh-access-token.usecase.spec.ts
@@ -1,15 +1,9 @@
 /* eslint-disable @typescript-eslint/unbound-method */
 import { createHash } from 'crypto';
 import { RefreshAccessTokenUseCase } from './refresh-access-token.usecase';
-import { AuthRepository } from '../../domain/auth.repository';
+import type { AuthSessionPort } from '../ports/auth-session.port';
 
-const createRepository = (): jest.Mocked<AuthRepository> => ({
-  findAccountByProvider: jest.fn(),
-  findAccountById: jest.fn(),
-  findUserById: jest.fn(),
-  findUserByAuthEmail: jest.fn(),
-  createUserWithAuthAccount: jest.fn(),
-  createAuthAccountForUser: jest.fn(),
+const createSessionPort = (): jest.Mocked<AuthSessionPort> => ({
   issueTokens: jest.fn(),
   signAccessToken: jest.fn(),
   findRefreshTokenSession: jest.fn(),
@@ -21,10 +15,10 @@ const createHashToken = (token: string): string =>
 
 describe('RefreshAccessTokenUseCase', () => {
   it('세션이 없으면 UNAUTHORIZED 오류를 반환한다', async () => {
-    const repo = createRepository();
-    repo.findRefreshTokenSession.mockResolvedValue(null);
+    const sessionPort = createSessionPort();
+    sessionPort.findRefreshTokenSession.mockResolvedValue(null);
 
-    const usecase = new RefreshAccessTokenUseCase(repo);
+    const usecase = new RefreshAccessTokenUseCase(sessionPort);
 
     await expect(
       usecase.execute(
@@ -35,14 +29,14 @@ describe('RefreshAccessTokenUseCase', () => {
   });
 
   it('폐기된 세션이면 UNAUTHORIZED 오류를 반환한다', async () => {
-    const repo = createRepository();
-    repo.findRefreshTokenSession.mockResolvedValue({
+    const sessionPort = createSessionPort();
+    sessionPort.findRefreshTokenSession.mockResolvedValue({
       tokenHash: createHashToken('refresh-token'),
       revokedAt: new Date(),
       expiresAt: new Date(Date.now() + 1000 * 60),
     });
 
-    const usecase = new RefreshAccessTokenUseCase(repo);
+    const usecase = new RefreshAccessTokenUseCase(sessionPort);
 
     await expect(
       usecase.execute(
@@ -53,14 +47,14 @@ describe('RefreshAccessTokenUseCase', () => {
   });
 
   it('만료된 세션이면 UNAUTHORIZED 오류를 반환한다', async () => {
-    const repo = createRepository();
-    repo.findRefreshTokenSession.mockResolvedValue({
+    const sessionPort = createSessionPort();
+    sessionPort.findRefreshTokenSession.mockResolvedValue({
       tokenHash: createHashToken('refresh-token'),
       revokedAt: null,
       expiresAt: new Date(Date.now() - 1000 * 60),
     });
 
-    const usecase = new RefreshAccessTokenUseCase(repo);
+    const usecase = new RefreshAccessTokenUseCase(sessionPort);
 
     await expect(
       usecase.execute(
@@ -71,14 +65,14 @@ describe('RefreshAccessTokenUseCase', () => {
   });
 
   it('토큰 해시가 다르면 UNAUTHORIZED 오류를 반환한다', async () => {
-    const repo = createRepository();
-    repo.findRefreshTokenSession.mockResolvedValue({
+    const sessionPort = createSessionPort();
+    sessionPort.findRefreshTokenSession.mockResolvedValue({
       tokenHash: createHashToken('stored-token'),
       revokedAt: null,
       expiresAt: new Date(Date.now() + 1000 * 60),
     });
 
-    const usecase = new RefreshAccessTokenUseCase(repo);
+    const usecase = new RefreshAccessTokenUseCase(sessionPort);
 
     await expect(
       usecase.execute(
@@ -89,21 +83,21 @@ describe('RefreshAccessTokenUseCase', () => {
   });
 
   it('정상적인 요청이면 액세스 토큰을 발급한다', async () => {
-    const repo = createRepository();
-    repo.findRefreshTokenSession.mockResolvedValue({
+    const sessionPort = createSessionPort();
+    sessionPort.findRefreshTokenSession.mockResolvedValue({
       tokenHash: createHashToken('refresh-token'),
       revokedAt: null,
       expiresAt: new Date(Date.now() + 1000 * 60),
     });
-    repo.signAccessToken.mockReturnValue('access-token');
+    sessionPort.signAccessToken.mockReturnValue('access-token');
 
-    const usecase = new RefreshAccessTokenUseCase(repo);
+    const usecase = new RefreshAccessTokenUseCase(sessionPort);
     const result = await usecase.execute(
       { userId: 'user-id', accountId: 'account-id', platform: 'WEB' },
       'refresh-token',
     );
 
-    expect(repo.signAccessToken).toHaveBeenCalledWith({
+    expect(sessionPort.signAccessToken).toHaveBeenCalledWith({
       userId: 'user-id',
       accountId: 'account-id',
       platform: 'WEB',
