+++ src/auth/application/usecases/refresh-access-token.usecase.ts
@@ -1,16 +1,16 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { createHash } from 'crypto';
-import { AUTH_REPOSITORY } from '../../domain/auth.repository';
-import type { AuthRepository } from '../../domain/auth.repository';
 import { AuthContext } from '../auth-context';
 import { AuthError } from '../auth.error';
+import { AUTH_SESSION_PORT } from '../ports/auth-session.port';
+import type { AuthSessionPort } from '../ports/auth-session.port';
 import { AccessTokenResult } from '../auth.types';
 
 @Injectable()
 export class RefreshAccessTokenUseCase {
   constructor(
-    @Inject(AUTH_REPOSITORY)
-    private readonly authRepository: AuthRepository,
+    @Inject(AUTH_SESSION_PORT)
+    private readonly authSessionPort: AuthSessionPort,
   ) {}
 
   async execute(
@@ -19,7 +19,7 @@ export class RefreshAccessTokenUseCase {
   ): Promise<AccessTokenResult> {
     const { accountId, platform } = context;
 
-    const session = await this.authRepository.findRefreshTokenSession(
+    const session = await this.authSessionPort.findRefreshTokenSession(
       accountId,
       platform,
     );
@@ -44,7 +44,7 @@ export class RefreshAccessTokenUseCase {
       throw new AuthError('UNAUTHORIZED', '리프레쉬 토큰이 일치하지 않습니다.');
     }
 
-    const accessToken = this.authRepository.signAccessToken(context);
+    const accessToken = this.authSessionPort.signAccessToken(context);
 
     return { access_token: accessToken };
   }
