+++ src/auth/application/usecases/switch-user.usecase.spec.ts
@@ -1,6 +1,7 @@
 /* eslint-disable @typescript-eslint/unbound-method */
 import { SwitchUserUseCase } from './switch-user.usecase';
-import { AuthRepository } from '../../domain/auth.repository';
+import type { AuthRepository } from '../../domain/auth.repository';
+import type { AuthSessionPort } from '../ports/auth-session.port';
 
 const createRepository = (): jest.Mocked<AuthRepository> => ({
   findAccountByProvider: jest.fn(),
@@ -9,6 +10,9 @@ const createRepository = (): jest.Mocked<AuthRepository> => ({
   findUserByAuthEmail: jest.fn(),
   createUserWithAuthAccount: jest.fn(),
   createAuthAccountForUser: jest.fn(),
+});
+
+const createSessionPort = (): jest.Mocked<AuthSessionPort> => ({
   issueTokens: jest.fn(),
   signAccessToken: jest.fn(),
   findRefreshTokenSession: jest.fn(),
@@ -18,9 +22,10 @@ const createRepository = (): jest.Mocked<AuthRepository> => ({
 describe('SwitchUserUseCase', () => {
   it('계정이 없으면 NOT_FOUND 오류를 반환한다', async () => {
     const repo = createRepository();
+    const sessionPort = createSessionPort();
     repo.findAccountById.mockResolvedValue(null);
 
-    const usecase = new SwitchUserUseCase(repo);
+    const usecase = new SwitchUserUseCase(repo, sessionPort);
 
     await expect(
       usecase.execute('user-id', 'account-id', 'WEB'),
@@ -29,6 +34,7 @@ describe('SwitchUserUseCase', () => {
 
   it('현재 사용자와 계정이 다르면 FORBIDDEN 오류를 반환한다', async () => {
     const repo = createRepository();
+    const sessionPort = createSessionPort();
     repo.findAccountById.mockResolvedValue({
       id: 'account-id',
       userId: 'other-user-id',
@@ -39,7 +45,7 @@ describe('SwitchUserUseCase', () => {
       profileImageUrl: null,
     });
 
-    const usecase = new SwitchUserUseCase(repo);
+    const usecase = new SwitchUserUseCase(repo, sessionPort);
 
     await expect(
       usecase.execute('user-id', 'account-id', 'WEB'),
@@ -48,6 +54,7 @@ describe('SwitchUserUseCase', () => {
 
   it('정상적인 요청이면 토큰을 발급한다', async () => {
     const repo = createRepository();
+    const sessionPort = createSessionPort();
     repo.findAccountById.mockResolvedValue({
       id: 'account-id',
       userId: 'user-id',
@@ -57,15 +64,15 @@ describe('SwitchUserUseCase', () => {
       displayName: 'Switch User',
       profileImageUrl: 'https://example.com/avatar.png',
     });
-    repo.issueTokens.mockResolvedValue({
+    sessionPort.issueTokens.mockResolvedValue({
       accessToken: 'access-token',
       refreshToken: 'refresh-token',
     });
 
-    const usecase = new SwitchUserUseCase(repo);
+    const usecase = new SwitchUserUseCase(repo, sessionPort);
     const result = await usecase.execute('user-id', 'account-id', 'WEB');
 
-    expect(repo.issueTokens).toHaveBeenCalledWith({
+    expect(sessionPort.issueTokens).toHaveBeenCalledWith({
       userId: 'user-id',
       accountId: 'account-id',
       platform: 'WEB',
