+++ src/auth/application/usecases/switch-user.usecase.ts
@@ -1,6 +1,8 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { AUTH_REPOSITORY } from '../../domain/auth.repository';
 import type { AuthRepository } from '../../domain/auth.repository';
+import { AUTH_SESSION_PORT } from '../ports/auth-session.port';
+import type { AuthSessionPort } from '../ports/auth-session.port';
 import { AuthError } from '../auth.error';
 import { OAuthSignInResult } from '../auth.types';
 import { AuthPlatform } from '../../domain/auth.types';
@@ -11,6 +13,8 @@ export class SwitchUserUseCase {
   constructor(
     @Inject(AUTH_REPOSITORY)
     private readonly authRepository: AuthRepository,
+    @Inject(AUTH_SESSION_PORT)
+    private readonly authSessionPort: AuthSessionPort,
   ) {}
 
   async execute(
@@ -29,7 +33,7 @@ export class SwitchUserUseCase {
       throw new AuthError('FORBIDDEN', '연동되지 않은 계정입니다.');
     }
 
-    return issueAuthResult(this.authRepository, {
+    return issueAuthResult(this.authSessionPort, {
       userId: targetAccount.userId,
       account: targetAccount,
       platform,
